#407: Moved the factory method for creating FtSecure fromt FtRemote to the test.
